+++ demo/server.php
@@ -3,13 +3,10 @@ declare( strict_types = 1 );
 
 require __DIR__ . '/../vendor/autoload.php';
 
-$router = new Taalkic\Router();
-require __DIR__ . '/url-routes.php';
-
 $app = new Taalkic\App( [
 	'workers' => 'half',
 	'port' => 4200,
-	'router' => $router,
+	'routes' => __DIR__ . '/url-routes.php',
 	'charset' => 'utf-8',
 	'template_dir' => __DIR__ . '/templates/',
 ] );